updates to product page to get variant selectors to work
this is a incremental fix where I am refreshing the variant selection on navigation. Its fixing the issue of stock display but it doesn't reset the variant button selections to the correct values.

this could work if all initial variants have the first values as default.

make sure to select the default values and images